+++ src/client/App.js
@@ -1,55 +1,12 @@
-import React, { useEffect, useState } from 'react';
+import React from 'react';
 import Post from './Post';
 import Form from './Form';
 import './App.css';
-const API = `/.netlify/functions`;
+import usePosts from './hooks/usePosts';
 
 // State for loading, error and posts
 const App = () => {
-  const [loading, setLoading] = useState(true);
-  const [error, setError] = useState(null);
-  const [posts, setPosts] = useState([]);
-
-  // When the component renders for the first time, fetch all the posts
-  useEffect(() => {
-    getPosts();
-  }, []);
-  async function getPosts() {
-    try {
-      setLoading(true);
-      const response = await fetch(`${API}/get-posts`);
-      const { posts, error } = await response.json();
-      if (error) throw new Error(error);
-      setPosts(posts);
-    } catch (error) {
-      setError(error);
-    } finally {
-      setLoading(false);
-    }
-  }
-
-  // Add or delete posts
-  async function addPost({ title, content }) {
-    try {
-      if (!title || !content) return;
-      setLoading(true);
-      const body = JSON.stringify({ title, content });
-      await fetch(`${API}/add-post`, { method: 'POST', body });
-      return getPosts(); // Refresh all posts
-    } catch (error) {
-      setError(error);
-    }
-  }
-  async function deletePost({ id }) {
-    try {
-      setLoading(true);
-      const body = JSON.stringify({ id });
-      await fetch(`${API}/delete-post`, { method: 'POST', body });
-      return getPosts(); // Refresh all posts
-    } catch (error) {
-      setError(error);
-    }
-  }
+  const { posts, loading, error, client } = usePosts();
 
   // If error or loading, show a message
   if (error) return <p>Error: {error.message}</p>;
@@ -59,10 +16,10 @@ const App = () => {
   return (
     <div className="App">
       <header>SimpleBlog</header>
-      <Form onAdd={({ title, content }) => addPost({ title, content })} />
+      <Form onAdd={({ title, content }) => client.add({ title, content })} />
       <div>
         {posts.map((post) => (
-          <Post data={post} onDelete={() => deletePost({ id: post.id })} />
+          <Post data={post} onDelete={() => client.delete({ id: post.id })} />
         ))}
       </div>
       <hr />
